+++ AirLiteTalonFX2/Core/Inc/quaternion.h
@@ -9,13 +9,14 @@
 #define INC_QUATERNION_H_
 
 #include "math.h"
+#include "arm_math.h"
 
-void Quaternion_Multiply(double *out, const double *q, const double *p);
+void Quaternion_Multiply(float32_t *out, const float32_t *q, const float32_t *p);
 
-void Quaternion_Normalize(double *q);
+void Quaternion_Normalize(float32_t *q);
 
-void Quaternion_To_Euler(const double quat[4], double euler[3]);
+void Quaternion_To_Euler(const float32_t quat[4], float32_t euler[3]);
 
-void Quaternion_To_Euler_Deg(const double quat[4], double euler[3]);
+void Quaternion_To_Euler_Deg(const float32_t quat[4], float32_t euler[3]);
 
 #endif /* INC_QUATERNION_H_ */
